+++ x10/perpetual/simple_client/simple_trading_client.py
@@ -4,11 +4,11 @@ import time
 from decimal import Decimal
 from typing import Awaitable, Dict, Union, cast
 
-from x10.perpetual.accounts import AccountStreamDataModel, StarkPerpetualAccount
-from x10.perpetual.configuration import EndpointConfig
-from x10.perpetual.markets import MarketModel
-from x10.perpetual.order_object import create_order_object
-from x10.perpetual.orders import (
+from x10.configuration import EndpointConfig
+from x10.core.stark_account import StarkPerpetualAccount
+from x10.models.account import AccountStreamDataModel
+from x10.models.market import MarketModel
+from x10.models.order import (
     NewOrderModel,
     OpenOrderModel,
     OrderSide,
@@ -16,6 +16,7 @@ from x10.perpetual.orders import (
     OrderType,
     TimeInForce,
 )
+from x10.perpetual.order_object import create_order_object
 from x10.perpetual.stream_client.perpetual_stream_connection import (
     PerpetualStreamConnection,
 )
