+++ x10/perpetual/user_client/onboarding.py
@@ -5,7 +5,6 @@ from eth_account.messages import SignableMessage, encode_typed_data
 from eth_account.signers.local import LocalAccount
 from fast_stark_crypto import generate_keypair_from_eth_signature, pedersen_hash
 from fast_stark_crypto import sign as stark_sign
-
 from models.account import AccountModel
 from models.base import X10BaseModel
